--- boot/src/main/java/com/rainyalley/architecture/config/GlobalExceptionHandler.java
@@ -1,21 +1,0 @@
-package com.rainyalley.architecture.config;
-
-import com.rainyalley.architecture.vo.ErrorVo;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(value = Exception.class)
-    @ResponseBody
-    public ResponseEntity<ErrorVo> defaultErrorHandler(Exception e) {
-        ErrorVo info = new ErrorVo();
-        info.setMessage(e.getMessage());
-        info.setCode(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(info);
-    }
-}
