+++ boot/src/main/java/com/rainyalley/architecture/filter/limit/DefaultStatisticsStrategy.java
@@ -1,4 +1,4 @@
-package com.rainyalley.architecture.filter;
+package com.rainyalley.architecture.filter.limit;
 
 import java.util.HashMap;
 import java.util.concurrent.atomic.AtomicInteger;
