+++ boot/src/main/java/com/rainyalley/architecture/filter/limit/GlobalLimitInfo.java
@@ -1,4 +1,4 @@
-package com.rainyalley.architecture.filter;
+package com.rainyalley.architecture.filter.limit;
 
 import java.util.Set;
