+++ boot/src/main/java/com/rainyalley/architecture/filter/limit/LimitStrategy.java
@@ -1,4 +1,4 @@
-package com.rainyalley.architecture.filter;
+package com.rainyalley.architecture.filter.limit;
 
 import javax.servlet.http.HttpServletRequest;
