+++ boot/src/main/java/com/rainyalley/architecture/filter/limit/StatisticsStrategy.java
@@ -1,4 +1,4 @@
-package com.rainyalley.architecture.filter;
+package com.rainyalley.architecture.filter.limit;
 
 public interface StatisticsStrategy {
