+++ pom.xml
@@ -251,6 +251,12 @@
                 <artifactId>jaxb-api</artifactId>
                 <version>2.3.0</version>
             </dependency>
+
+            <dependency>
+                <groupId>org.jsoup</groupId>
+                <artifactId>jsoup</artifactId>
+                <version>1.11.2</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
